# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: web-mini-template API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/tasks:
    post:
      operationId: createTask
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Created
components:
  schemas:
...
    CreateTaskRequest:
      type: object
      required:
        - title
        - member_id
      properties:
        title:
          type: string
        member_id:
          type: integer
          format: int64