# server/internal/gen/openapi/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package openapi

// CreateTaskCreated is response for CreateTask operation.
type CreateTaskCreated struct{}

// Ref: #/components/schemas/CreateTaskRequest
type CreateTaskRequest struct {
	Title    string `json:"title"`
	MemberID int64  `json:"member_id"`
}

// GetTitle returns the value of Title.
func (s *CreateTaskRequest) GetTitle() string {
	return s.Title
}

// GetMemberID returns the value of MemberID.
func (s *CreateTaskRequest) GetMemberID() int64 {
	return s.MemberID
}

// SetTitle sets the value of Title.
func (s *CreateTaskRequest) SetTitle(val string) {
	s.Title = val
}

// SetMemberID sets the value of MemberID.
func (s *CreateTaskRequest) SetMemberID(val int64) {
	s.MemberID = val
}

// Ref: #/components/schemas/FeedItem
type FeedItem struct {
	ProjectID   int64  `json:"project_id"`
	ProjectName string `json:"project_name"`
	OwnerName   string `json:"owner_name"`
	TaskID      int64  `json:"task_id"`
	TaskTitle   string `json:"task_title"`
	TaskStatus  string `json:"task_status"`
}

// GetProjectID returns the value of ProjectID.
func (s *FeedItem) GetProjectID() int64 {
	return s.ProjectID
}

// GetProjectName returns the value of ProjectName.
func (s *FeedItem) GetProjectName() string {
	return s.ProjectName
}

// GetOwnerName returns the value of OwnerName.
func (s *FeedItem) GetOwnerName() string {
	return s.OwnerName
}

// GetTaskID returns the value of TaskID.
func (s *FeedItem) GetTaskID() int64 {
	return s.TaskID
}

// GetTaskTitle returns the value of TaskTitle.
func (s *FeedItem) GetTaskTitle() string {
	return s.TaskTitle
}

// GetTaskStatus returns the value of TaskStatus.
func (s *FeedItem) GetTaskStatus() string {
	return s.TaskStatus
}

// SetProjectID sets the value of ProjectID.
func (s *FeedItem) SetProjectID(val int64) {
	s.ProjectID = val
}

// SetProjectName sets the value of ProjectName.
func (s *FeedItem) SetProjectName(val string) {
	s.ProjectName = val
}

// SetOwnerName sets the value of OwnerName.
func (s *FeedItem) SetOwnerName(val string) {
	s.OwnerName = val
}

// SetTaskID sets the value of TaskID.
func (s *FeedItem) SetTaskID(val int64) {
	s.TaskID = val
}

// SetTaskTitle sets the value of TaskTitle.
func (s *FeedItem) SetTaskTitle(val string) {
	s.TaskTitle = val
}

// SetTaskStatus sets the value of TaskStatus.
func (s *FeedItem) SetTaskStatus(val string) {
	s.TaskStatus = val
}

// Ref: #/components/schemas/FeedResponse
type FeedResponse struct {
	Data []FeedItem `json:"data"`
}

// GetData returns the value of Data.
func (s *FeedResponse) GetData() []FeedItem {
	return s.Data
}

// SetData sets the value of Data.
func (s *FeedResponse) SetData(val []FeedItem) {
	s.Data = val
}

// InitializeNoContent is response for Initialize operation.
type InitializeNoContent struct{}

// Ref: #/components/schemas/Member
type Member struct {
	ID   int64  `json:"id"`
	Name string `json:"name"`
}

// GetID returns the value of ID.
func (s *Member) GetID() int64 {
	return s.ID
}

// GetName returns the value of Name.
func (s *Member) GetName() string {
	return s.Name
}

// SetID sets the value of ID.
func (s *Member) SetID(val int64) {
	s.ID = val
}

// SetName sets the value of Name.
func (s *Member) SetName(val string) {
	s.Name = val
}

// Ref: #/components/schemas/MembersResponse
type MembersResponse struct {
	Data []Member `json:"data"`
}

// GetData returns the value of Data.
func (s *MembersResponse) GetData() []Member {
	return s.Data
}

// SetData sets the value of Data.
func (s *MembersResponse) SetData(val []Member) {
	s.Data = val
}